third_party/WebKit/LayoutTests/fast/events/pointerevents/mouse-pointer-updown-events.html

<!DOCTYPE HTML>
<script src="../../../resources/js-test.js"></script>
<style>
div.box {
  margin: 10px;
  padding: 50px;
  float: left;
}
</style>

<div id="lightgreen" class="box" style="background-color:lightgreen">
  <div id="green" class="box" style="background-color:green">
  </div>
</div>

<div id="console"></div>

<script>
description("Verifies that pointerup/down are fired correctly for corresponding mouse events.");

function init() {
  var eventList = ["mousedown", "mouseup", "pointerdown", "pointerup"];

  ["green", "lightgreen"].forEach(function(id) {
    var targetDiv = document.getElementById(id);

    eventList.forEach(function(eventName) {
      targetDiv.addEventListener(eventName, function(event) {
        debug(id + " received " + event.type);
      });
    });
  });
}

function runTests() {
  var rect = document.getElementById("green").getBoundingClientRect();
  var x1 = rect.left + 5;
  var y1 = rect.top + 5;

  var rect = document.getElementById("lightgreen").getBoundingClientRect();
  var x2 = rect.left + 5;
  var y2 = rect.top + 5;

  eventSender.mouseMoveTo(x1, y1);
  eventSender.mouseDown();
  eventSender.mouseUp();
  eventSender.mouseDown();

  eventSender.mouseMoveTo(x2, y2);
  eventSender.mouseUp();
  eventSender.mouseDown();

  eventSender.mouseMoveTo(x1, y1);
  eventSender.mouseUp();
}

init();
if (window.eventSender)
  runTests();
else
  debug("This test requires eventSender");

</script>
