third_party/WebKit/LayoutTests/fast/events/pointerevents/mouse-pointer-updown-events-expected.txt

Verifies that pointerup/down are fired correctly for corresponding mouse events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


green received pointerdown
lightgreen received pointerdown
green received mousedown
lightgreen received mousedown
green received pointerup
lightgreen received pointerup
green received mouseup
lightgreen received mouseup
green received pointerdown
lightgreen received pointerdown
green received mousedown
lightgreen received mousedown
lightgreen received pointerup
lightgreen received mouseup
lightgreen received pointerdown
lightgreen received mousedown
green received pointerup
lightgreen received pointerup
green received mouseup
lightgreen received mouseup
PASS successfullyParsed is true

TEST COMPLETE
